chore(epic-5): final polish sweep — 4 items + Move B handler narrowing
Closes the 4 remaining Epic 5 deferred items in one cycle:

1. LoyaltyAccrualEntity.points INT→BIGINT (V008): widens the column to
   prevent theoretical 32-bit overflow. Use case + saga + controller +
   tests updated (int→long cascade).

2. AppendOrderTransitionCommand.pointsApplied: nullable Long field +
   backward-compat 4-arg ctor (existing callers default to null). Future
   RedeemLoyaltyPointsUseCase will read it.

3. Pricebook.load() fail-fast: throws IllegalStateException on
   missing/malformed pricebook.json (was log+swallow). ADR-20 parity.

4. PricingController @WebMvcTest: 2 cases (200 known + 404 unknown)
   via MockMvcBuilders.standaloneSetup. Pricebook 6/6 (was 3, +3).

Plus: Move B's broad IllegalArgumentException handler in
OrderEditExceptionHandler was replaced with a targeted
OrderSnapshotMissingException handler. Root-cause fix —
OrderController.accrueLoyalty now throws the dedicated type
(already existed from Story 4.2). IAE from any other controller in
the order module no longer maps to 404.

Order 68/68 (unchanged), pricing 6/6, util 78/78, auth 28/28.

+++ services/order/src/main/java/vn/vnpt/order/application/saga/PaymentCapturedOrderAdvancer.java
@@ -110,7 +110,7 @@ public class PaymentCapturedOrderAdvancer {
           // v1: snapshot.userId IS the customerId (auth's users.customer_id == userId).
           // Forward-compat: replace with a user→customer cross-service lookup if the key
           // model diverges.
-          int accrued = accrueLoyaltyPointsUseCase.execute(event.orderUuid(), userId, snapshot.getTotalCents());
+          long accrued = accrueLoyaltyPointsUseCase.execute(event.orderUuid(), userId, snapshot.getTotalCents());
           if (accrued == 0) {
             log.debug("Loyalty accrual returned 0 points for orderUuid={} (customerId={})",
                 event.orderUuid(), userId);
